fix:update identifier implementation
