+++ src/main/java/org/eclipse/tractusx/managedidentitywallets/controller/DidDocumentController.java
@@ -21,8 +21,6 @@
 
 package org.eclipse.tractusx.managedidentitywallets.controller;
 
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
